refactor: 패키지 구조 변경 및 Executor 추상화
