+++ app/src/main/java/com/techcourse/service/impl/TxUserService.java
@@ -1,17 +1,18 @@
-package com.techcourse.service;
+package com.techcourse.service.impl;
 
 import com.techcourse.config.DataSourceConfig;
 import com.techcourse.domain.User;
-import org.springframework.transaction.support.TransactionExecutor;
+import com.techcourse.service.UserService;
+import org.springframework.transaction.support.ServiceExecutor;
 
 public class TxUserService implements UserService {
 
     private final UserService userService;
-    private final TransactionExecutor transactionExecutor;
+    private final ServiceExecutor serviceExecutor;
 
     public TxUserService(final UserService userService) {
         this.userService = userService;
-        this.transactionExecutor = new TransactionExecutor(DataSourceConfig.getInstance());
+        this.serviceExecutor = new ServiceExecutor(DataSourceConfig.getInstance());
     }
 
     @Override
@@ -26,7 +27,7 @@ public class TxUserService implements UserService {
 
     @Override
     public void changePassword(final long id, final String newPassword, final String createBy) {
-        transactionExecutor.execute(() -> {
+        serviceExecutor.execute(() -> {
             userService.changePassword(id, newPassword, createBy);
             return null;
         });
