+++ jdbc/src/main/java/org/springframework/transaction/support/TransactionExecutor.java
@@ -1,41 +1,6 @@
 package org.springframework.transaction.support;
 
-import java.sql.Connection;
-import java.sql.SQLException;
-import java.util.function.Supplier;
-import javax.sql.DataSource;
-import org.springframework.dao.DataAccessException;
-import org.springframework.jdbc.datasource.DataSourceUtils;
-
-public class TransactionExecutor {
-
-    private final DataSource dataSource;
-
-    public TransactionExecutor(DataSource dataSource) {
-        this.dataSource = dataSource;
-    }
-
-    public <T> T execute(Supplier<T> supplier) {
-        final Connection conn = DataSourceUtils.getConnection(dataSource);
-        try {
-            conn.setAutoCommit(false);
-            final T result = supplier.get();
-            conn.commit();
-            return result;
-        } catch (Exception e) {
-            rollback(conn);
-            throw new DataAccessException(e);
-        } finally {
-            DataSourceUtils.releaseConnection(conn, dataSource);
-            TransactionSynchronizationManager.unbindResource(dataSource);
-        }
-    }
-
-    private void rollback(final Connection conn) {
-        try {
-            conn.rollback();
-        } catch (SQLException exception) {
-            throw new DataAccessException(exception);
-        }
-    }
+@FunctionalInterface
+public interface TransactionExecutor<T> {
+    T execute();
 }
